Fixes issue #1
Move all event receiving code to the activity from the fragments.
